refactor: move out hotel routes from index to routes folder
